% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{allplots}
\alias{allplots}
\title{Show all plots in a result object}
\usage{
allplots(res)
}
\arguments{
\item{res}{A result object}
}
\value{
all the plots (in a list, but they display by default)
}
\description{
Show all plots in a result object
}

