% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Anchovy-class}
\alias{Anchovy-class}
\title{An S4 class for Anchovy.}
\value{
An object of the Anchovy class.
}
\description{
Inherits from Prey.
}
\section{Slots}{

\describe{
\item{\code{p_fail}}{The probability of recruitment failure}

\item{\code{R_residuals}}{A set of observed recruitment residuals.}
}}
\examples{
anch<-new("Anchovy")
slotNames(anch) #see list of slots
anch@Zmat[,1:10] #show total mortality by age (Zmat) for timesteps 1 to 10
}
\seealso{
parent classes \linkS4class{Prey} and \linkS4class{Species} (top-level).
}

