% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_availability}
\alias{calc_availability}
\title{Calculate prey Availability parameter}
\usage{
calc_availability(ap, ac, Bt)
}
\arguments{
\item{ap}{The p parameter of a Beverton-Holt function}

\item{ac}{The c parameter of a Beverton-Holt function}

\item{Bt}{The biomass of prey at time t}
}
\value{
A value that represents some fraction of Bt.
}
\description{
The availability function is a surrogate for the spatial availability of prey
to predators (i.e., it answers the question "How much of the prey population
can the predators access?")
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}}, \code{\link{calc_other_R}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

