% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_B1P}
\alias{calc_B1P}
\title{Calculate age 1+ biomass}
\usage{
calc_B1P(Nv, wv)
}
\arguments{
\item{Nv}{Numbers-at-age}

\item{wv}{Weight-at-age}
}
\value{
A scalar of total spawning biomass of fish aged 1+ (i.e., excludes age-0), in tons ('000 kg).
}
\description{
Used for calculating prey biomass of a species in one timestep or year, where
prey size matters to the predator but prey maturity does not).
}
\seealso{
Other recruitment: \code{\link{calc_B_sp}},
  \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

