% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_B_sp}
\alias{calc_B_sp}
\title{Calculate spawning biomass (tons)}
\usage{
calc_B_sp(wv, Nv, maturev)
}
\arguments{
\item{wv}{Weight at age (in kg)}

\item{Nv}{Numbers at age}

\item{maturev}{maturity at age}
}
\value{
A scalar: biomass in tons
}
\description{
A utility function used in recruitment calculations, in some species.
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}}, \code{\link{calc_phi}},
  \code{\link{calc_predRecruits_Punt}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

