% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_Bev_Holt}
\alias{calc_Bev_Holt}
\title{Calculate recruits, using a Beverton-Holt model}
\usage{
calc_Bev_Holt(Nm, p, c)
}
\arguments{
\item{Nm}{Number of mature predators.}

\item{p}{A parameter that defines maximum offspring per predator at low densities.}

\item{c}{The parameter that defines carrying capacity.}
}
\value{
Number of recruits (scalar).
}
\description{
Calculate recruits, using a Beverton-Holt model
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

