% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_Bprey_agepreferred}
\alias{calc_Bprey_agepreferred}
\title{Calculate the age-preferred biomass of a prey species available to a predator}
\usage{
calc_Bprey_agepreferred(predator, prey, t)
}
\arguments{
\item{predator}{A predator object.}

\item{prey}{A prey object.}

\item{t}{Time t.}
}
\value{
A biomass value.
}
\description{
A utility function that calculates the biomass (tons) of prey species X that
is available to predator species Y, as:  (biomass-at-age of prey) x (predator
preferences for prey age classes). This function returns an intermediate
product in the calculation of total available prey biomass. Note: this is
akin to Punt et al's use of biomass of age 1+ fish, but allows the predator
to eat some of the age-0 fish as well if their age-preferences indicate it.
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

