% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_catch}
\alias{calc_catch}
\title{Calculate catch (biomass units)}
\usage{
calc_catch(Nv, wv, Mv, Sv, Ft)
}
\arguments{
\item{Nv}{Vector of numbers at age for a particular time t}

\item{wv}{Vector of weight at age}

\item{Mv}{Vector of natural mortality rate at age}

\item{Sv}{Vector of fishing gear selectivity at age}

\item{Ft}{Instantaneous fishing mortality rate (log scale); a scalar.}

\item{Species}{A Species object}

\item{t}{time period (year or timestep)}
}
\value{
Catch in biomass units.
}
\description{
A utility function for calculating catch at time t.  Uses the Baranov equation,
with fishing selectivity by age.
}
\seealso{
Other fishing: \code{\link{calc_Ft}}, \code{\link{calc_Fy}}
}
\keyword{internal}

