% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_Dy}
\alias{calc_Dy}
\title{Calculate Dy.}
\usage{
calc_Dy(predator, B0list, preylist)
}
\arguments{
\item{predator}{A Predator object.}

\item{B0list}{A named list of available prey biomass for each species.}

\item{preylist}{A list of Prey objects.}
}
\value{
A value for Dy (scalar).
}
\description{
Calculates the scaled, total biomass of all prey, Dy, adjusted for species-
and age-class preferences and spatial availability, for a predator X.  Dy is
set so that Dy = 1 and mean diet proportions of prey species in the diet of
predator X = observed diet proportions when the prey are unfished.
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_R_error}}, \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}}, \code{\link{calc_phi}},
  \code{\link{calc_predRecruits_Punt}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

