% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_Ft}
\alias{calc_Ft}
\title{Calculate fishing mortality according to a fishing rule}
\usage{
calc_Ft(Species, y, ts, t)
}
\arguments{
\item{Species}{A Species object.}

\item{y}{The simulation year.}

\item{ts}{The global timestep ()}

\item{t}{The intra-year timestep.}
}
\value{
A scalar, Ft, the total fishing mortality rate (log scale) for timestep t.
}
\description{
There are two versions: one for species tracked by year, the other for species
tracked by timestep. Note: it would be easy to make this a generic and set up
alternative models for each species. This version is for species that are
tracked by timestep.  ts= t + (y * t).
}
\seealso{
Other fishing: \code{\link{calc_Fy}},
  \code{\link{calc_catch}}
}
\keyword{internal}

