% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_Fy}
\alias{calc_Fy}
\title{Calculate fishing mortality according to a fishing rule}
\usage{
calc_Fy(Species, y)
}
\arguments{
\item{Species}{A Species object.}

\item{y}{The current year.}

\item{t}{The intra-year timestep.}
}
\value{
A scalar, Fv: the total fishing mortality rate year (log scale) for year y.
}
\description{
This version is for species tracked by year.  Note: it would be easy to make
this a generic and set up alternative models for each species.
}
\seealso{
Other fishing: \code{\link{calc_Ft}},
  \code{\link{calc_catch}}
}
\keyword{internal}

