% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_mortality_t}
\alias{calc_mortality_t}
\title{Calculate total mortality rate}
\usage{
calc_mortality_t(M, L, Ft, Sv)
}
\arguments{
\item{M}{Natural mortality rate (scalar).}

\item{L}{Number of timesteps per year (scalar).}

\item{Ft}{Total fishing mortality rate (scalar).}

\item{Sv}{Fishing selectivity (vector).}
}
\value{
A vector of mortality rate (log scale) by age.
}
\description{
A utility function used to calculate total mortality rate by age in species
that are tracked by timestep.
}
\keyword{internal}

