% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_mortality_y}
\alias{calc_mortality_y}
\title{Calculate total mortality rate}
\usage{
calc_mortality_y(Mv, Ft, Sv)
}
\arguments{
\item{Mv}{Natural mortality rate (log scale) by age. Vector.}

\item{Ft}{Total fishing mortality rate (scalar).}

\item{Sv}{Fishing selectivity (vector).}
}
\value{
A vector of mortality rate (log scale) by age.
}
\description{
A utility function used to calculate total mortality rate by age in species
that are tracked by year.
}
\keyword{internal}

