% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_Ny}
\alias{calc_Ny}
\title{Calculate Numbers at age for year y, for species tracked by year.}
\usage{
calc_Ny(Species, y)
}
\arguments{
\item{Species}{A Species object.}

\item{y}{The current year.}
}
\value{
A modified Species object; see Description.
}
\description{
Calculates numbers at age in year y, Nmat[,y], from Nmat[,y-1]. Sets the
following values in the Species object: Nmat[,y]; the previous year's total
mortality rate (log scale) by age (Zmat[,y-1]); the current year's
fishing mortality rate (Fv[y], log scale, a scalar, fished species only) and
the current year age 1+ biomass, B1P[y]
(prey species only).
}
\seealso{
Other Basic_dynamics: \code{\link{calc_Nt}}
}
\keyword{internal}

