% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_omegas}
\alias{calc_omegas}
\title{Do a calibration run.}
\usage{
calc_omegas(predator, prey, nYears, nSteps)
}
\arguments{
\item{predator}{A list of Predator objects}

\item{prey}{A list of Prey objects}

\item{nYears}{Number of years for the simulation}

\item{nSteps}{Number of timesteps per year.}
}
\value{
A modified list of predators in which the @omegas slot, which holds
  scaling parameters for the ratio of diet preference to unfished biomass for
  each prey species, has been set for each predator.  NOTE: the prey list is
  unchanged.
}
\description{
Do an initial run for prey species without fishing, to calculate parameters
in the Predator objects (omegas) that scale the unfished biomass of each prey species.
}
\seealso{
Other initialization: \code{\link{distribute_byage}},
  \code{\link{distribute_bytimestep}},
  \code{\link{initialize_N}}, \code{\link{set_simParam}}
}
\keyword{internal}

