% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_other_R}
\alias{calc_other_R}
\title{Calculate recruits for Other_forage species}
\usage{
calc_other_R(Other, y)
}
\arguments{
\item{Other}{An Other_forage species object.}

\item{y}{The current year (integer).}
}
\value{
A modified Other_forage object, with recruitment set both in the population matrix (Nmat[1,y]) and in the R slot (R[y]).
}
\description{
As implemented, recruitment is constant at 10,000; however, log-scale
recruitment error is added as an autocorrelated random draw from a log-normal
distribution.
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

