% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_p}
\alias{calc_p}
\title{Calculate impact of prey biomass on predator reproduction}
\usage{
calc_p(predator, Dy)
}
\arguments{
\item{predator}{A Predator object}

\item{Dy}{The total biomass of prey available to the predator.}
}
\value{
The Beverton-Holt p parameter value (a scalar), scaled to represent some
  fraction of the maximum offspring per predator.
}
\description{
Calculates the impact of available prey biomass on the maximum offspring per
predator, for predator species X.  The function has two parameters: K, which
controls how steep the logistic inflection is, and Dhat, which controls the
location of the inflection point on the X axis.  Those parameters and Mo, the
maximum offspring per spawner, are stored in the Predator object.
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}}, \code{\link{calc_phi}},
  \code{\link{calc_predRecruits_Punt}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

