% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_phi}
\alias{calc_phi}
\title{Calculate phi (impact of prey on predator reproduction)}
\usage{
calc_phi(predator, Dy)
}
\arguments{
\item{predator}{A Predator object}

\item{Dy}{The relative (scaled) biomass of prey available to the predator.
Note: we're using one set of thetas for all prey species, but we could
potentially make them different for each prey (Punt et al. set them all equal
in the base scenario.)}
}
\value{
The parameter phi
}
\description{
Calculates the impact of prey on predator reproduction; see equation 12 of
Punt et al. 2016
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}},
  \code{\link{calc_predRecruits_Punt}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

