% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_predatorRecruits}
\alias{calc_predatorRecruits}
\title{Calculate Predator Recruits}
\usage{
calc_predatorRecruits(Predator, preylist, y, ts, nSteps)
}
\arguments{
\item{Predator}{A Predator object.}

\item{preylist}{A list containing the Prey objects.}

\item{y}{The current year.}

\item{ts}{The current global timestep.}

\item{nSteps}{The number of timesteps per year.}
}
\value{
A modified Predator object, with number of mature predators (Nm),
  predator recruits (Nmat[,t]) both set.  As a convience for later graphing,
  the number of recruits is also set in the predator's @R slot, and Dy (total,
  scaled prey biomass) and BHp (Beverton-Holt maximum growth parameter) are also set.
}
\description{
This function calls several utility functions to calculate predator recruits.
In brief: 1) it calculates the number of breeding predator adults; 2) it
calculates prey biomass; 3) it calculates the proportion of the prey biomass
that is available to the predator, given a) the predator's diet preferences
for prey species, b) the predator's preferences for certain age classes of
those species, and c) the predator's ability to access the prey (spatial
availability).
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

