% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{calc_predicted}
\alias{calc_predicted}
\title{Calculate predicted values}
\usage{
calc_predicted(species, predicted)
}
\arguments{
\item{species}{A species object (whatever is being plotted)}

\item{predicted}{Character; the name of a predicted series}
}
\value{
A data frame of predicted data (x and y; 100 rows)
}
\description{
A utility function for plotting, which calculates various kinds of predicted curves
for overlay in plots.
}
\seealso{
Other results: \code{\link{compile_results}},
  \code{\link{makeplot}}, \code{\link{plot_every_nth}},
  \code{\link{plot_fitted}},
  \code{\link{plot_function_only}},
  \code{\link{plot_summed_by_year}}
}
\keyword{internal}

