% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_predRecruits_Punt}
\alias{calc_predRecruits_Punt}
\title{Calculate recruits using Punt method}
\usage{
calc_predRecruits_Punt(predator, Nm, phi, N1P, K1P)
}
\arguments{
\item{predator}{A Predator object}

\item{Nm}{Number of mature predators (a scalar)}

\item{phi}{The prey impact parameter}
}
\value{
The number of recruits (a scalar)
}
\description{
Calculates recruits using line 1 of equation 9 in Punt et al. 2016.
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}}, \code{\link{calc_phi}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

