% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_preyRecruits}
\alias{calc_preyRecruits}
\title{Calculate Prey recruits}
\usage{
calc_preyRecruits(Prey, preylist, ts, y)
}
\arguments{
\item{Prey}{A prey object.}

\item{preylist}{A list of prey.  Used when the reproduction of one prey
species depends on another (in this implementation, anchovy recruitment
depends on the biomass of sardines, since sardines eat anchovies).}

\item{ts}{The current global timestep (ts = t + y * t; where t is the intra-year timestep)}

\item{y}{The current year (integer).}
}
\value{
A modified Prey object, with recruits for the year set in the
  population matrix Nmat[1,t] (t may be a year or timestep).  Also sets R in
  R[t] and spawning biomass in B_sp[t], for convenience in graphing.
}
\description{
Calculate prey recruits at one timestep or year. As implemented, it
calculates the *number* of recruits from the *biomass* of spawners.  In
brief: 1) it calculates spawning biomass, 2) in the case of anchovy, it
calculates the biomass of sardines and adjusts the recruitment model
parameters in response, 3) it chooses a value of log recruitment error for
the year or timestep, and 4) calculates recruitment, using a Ricker model
with an environmental driver.
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_p}}, \code{\link{calc_recruits}}
}
\keyword{internal}

