% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_R_error}
\alias{calc_R_error}
\title{Calculate stochastic recruitment errors}
\usage{
calc_R_error(Species)
}
\arguments{
\item{Species}{}
}
\value{
A modified Species object, with recruitment errors (1 x nTimes)
  vector stored in a slot called lnSR_err
}
\description{
Calculate stochastic recruitment errors (either autocorrelated or not) for
all nYears for a Species.  Sigma and rho vary by species.  In this version,
the errors are drawn from a log-normal distribution
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}}, \code{\link{calc_phi}},
  \code{\link{calc_predRecruits_Punt}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

