% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\docType{methods}
\name{calc_recruits}
\alias{calc_recruits}
\alias{calc_recruits,Other_forage,ANY-method}
\alias{calc_recruits,Predator,list-method}
\alias{calc_recruits,Prey,ANY-method}
\title{Calculate recruits.}
\usage{
calc_recruits(species, preylist, ...)

\S4method{calc_recruits}{Prey,ANY}(species, preylist, ts, y, ...)

\S4method{calc_recruits}{Other_forage,ANY}(species, y, ...)

\S4method{calc_recruits}{Predator,list}(species, preylist, y, ts, nSteps, ...)
}
\arguments{
\item{...}{Other arguments (from child methods).}

\item{species.}{A Species object.}

\item{preylist.}{A list of Prey objects (only used for Predators).}
}
\value{
A child method, depending on signature.
}
\description{
Calculate recruits.

The default method calculates recruits in a Prey object.

The Other_forage method calculates recruits for Other_forage species.

The Predator method calculates recruits for Predators.
}
\section{Methods (by class)}{
\itemize{
\item \code{species = Prey,preylist = ANY}: Prey method.

\item \code{species = Other_forage,preylist = ANY}: Other_forage method.

\item \code{species = Predator,preylist = list}: Predator method.
}}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_Ricker}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}}, \code{\link{calc_phi}},
  \code{\link{calc_predRecruits_Punt}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}}
}
\keyword{internal}

