% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_Ricker}
\alias{calc_Ricker}
\title{Calculate recruits per spawner (Ricker method)}
\usage{
calc_Ricker(alpha, beta, B_sp)
}
\arguments{
\item{alpha}{Ricker alpha.}

\item{beta}{Ricker beta.}

\item{B_sp}{Spawning biomass (in tons)}
}
\value{
Recruits (a scalar).
}
\description{
Note expansion by 1000 for this case.
}
\seealso{
Other recruitment: \code{\link{calc_B1P}},
  \code{\link{calc_B_sp}}, \code{\link{calc_Bev_Holt}},
  \code{\link{calc_Bprey_agepreferred}},
  \code{\link{calc_Dy}}, \code{\link{calc_R_error}},
  \code{\link{calc_availability}},
  \code{\link{calc_other_R}},
  \code{\link{calc_predatorRecruits}},
  \code{\link{calc_preyRecruits}}, \code{\link{calc_p}},
  \code{\link{calc_recruits}}
}
\keyword{internal}

