% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calibrate}
\alias{calibrate}
\title{Do a calibration run.}
\usage{
calibrate(predators, prey, nYears, nSteps)
}
\arguments{
\item{predators}{A list of Predator objects}

\item{prey}{A list of Prey objects}

\item{nYears}{Number of years for the simulation}

\item{nSteps}{Number of timesteps per year.}
}
\value{
A modified list of predators and prey.  In the predators, the @omegas
  slot, which holds scaling parameters for the ratio of diet preference to
  unfished biomass for each prey species, has been set.  The function also
  sets D0, the scaled, available prey biomass at unfished levels.  In the
  prey objects, the age-1+ unfished biomass, B0, is set.
}
\description{
Do an initial run for prey species without fishing, to calculate parameters
in the Predator objects (omegas) that scale the unfished biomass of each prey species.
}
\seealso{
Other initialization: \code{\link{distribute_byage}},
  \code{\link{distribute_bytimestep}},
  \code{\link{initialize_N}}, \code{\link{set_simParam}},
  \code{\link{solve_age0_survival}}
}
\keyword{internal}

