% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{get_every_nth}
\alias{get_every_nth}
\title{Get every nth value of a time series}
\usage{
get_every_nth(nSteps, pdata, xname, yname)
}
\arguments{
\item{nSteps}{The number of timesteps per year}

\item{pdata}{A small data frame with the by-timestep series in it}

\item{xname}{The name of the x series}

\item{yname}{The name of the y series}
}
\value{
A data frame (x,y) of every nth value of a time series
}
\description{
A utility function that returns every nth intra-annual step from a time
series that includes intra-annual steps.
}
\seealso{
Other results: \code{\link{calc_predicted}},
  \code{\link{compile_results}}, \code{\link{makeplot}},
  \code{\link{plot_fitted_multipanel}},
  \code{\link{plot_fitted}},
  \code{\link{plot_function_only}},
  \code{\link{plot_summed_by_year}}
}
\keyword{internal}

