% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\name{get_params}
\alias{get_params}
\title{Set parameter values in a Species object}
\usage{
get_params(Species, simParams)
}
\arguments{
\item{simParams}{A simulationParams object containing input data.}

\item{Species.}{A Species object.}
}
\value{
A modified Species object, with parameters set from the input data.
}
\description{
Reads values from a simulationParams class object and sets them
(appropriately customized for each species) in a Species-class object.  The
parameters for certain models are encapsulated in lists, which makes it
easier to change them later.
}
\seealso{
\code{\link{get_prefs}} for reading in diet preferences,  and
  \code{\link{set_species}} for calling both functions.

Other parameter setting: \code{\link{get_prefs}},
  \code{\link{initialize_N}}, \code{\link{set_simParam}},
  \code{\link{set_species}}
}
\keyword{internal}

