% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\name{get_prefs}
\alias{get_prefs}
\title{Set diet preference slots in a Predator object.}
\usage{
get_prefs(Predator, simParams)
}
\arguments{
\item{Predator}{A Predator object.}

\item{simParams}{A simulationParams object containing input data.}
}
\value{
A modified Predator object, with diet preference slots filled.
}
\description{
The function reads values from a simulationParams-class object and fills the
dietary preference slots in the Predator object.
}
\seealso{
\code{\link{get_params}} for reading in parameter values and
  \code{\link{set_species}} for calling both functions.

Other parameter setting: \code{\link{get_params}},
  \code{\link{initialize_N}}, \code{\link{set_simParam}},
  \code{\link{set_species}}
}
\keyword{internal}

