% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\name{initialize_N}
\alias{initialize_N}
\title{Initialize a Species object.}
\usage{
initialize_N(Species, simulationParams)
}
\arguments{
\item{Species.}{A species object.}

\item{simulationParams.}{A simulationParams object that holds global values and input data.}
}
\value{
A modified Species object, with arrays initialized.
}
\description{
This function creates the population matrix, Nmat, for a Species, whose dimension is
either (nAges x nYears) if use_timestep is FALSE, or (nAges x (nYears *
nSteps)) if use_timestep is TRUE.  It fills Nmat[,year==1], based on values in the
simulationParams object.  It also dimensions empty vectors that will hold the
value of recruitment residuals, fishing mortality and catch (for fished
species), age 1+ biomass (for prey) and proportion mature-at-age, and calculates
Nm[1], and sets R[1] to zero.
}
\seealso{
Other initialization: \code{\link{calc_omegas}},
  \code{\link{distribute_byage}},
  \code{\link{distribute_bytimestep}},
  \code{\link{set_simParam}}

Other parameter setting: \code{\link{get_params}},
  \code{\link{get_prefs}}, \code{\link{set_simParam}},
  \code{\link{set_species}}
}
\keyword{internal}

