% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{justplot}
\alias{justplot}
\title{Show one type of plot for all species}
\usage{
justplot(res, plotname)
}
\arguments{
\item{res}{Results object}

\item{plotname}{The name of a plot type (e.g., "B1P" or "R")}
}
\value{
Returns a list of plots (also displays them by default in RStudio)
}
\description{
Show one type of plot for all species
}
\examples{
\dontrun{
  justplot(res,"R") #plots the "R" (recruitment) plots for each species
  #To find plotnames, look at the plotnames for the species you're interested
  in, for example;
  names(res$Anchovy$plots)
  #or, to see what a particular plot looks like:
  res$Anchovy$plots$R
}
}

