% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_initialization.R
\name{load_residuals}
\alias{load_residuals}
\title{Load anchovy recruitment residuals}
\usage{
load_residuals(anchovy, simParams)
}
\arguments{
\item{anchovy}{An Anchovy object.}

\item{simParams}{A simulationParameters object.}
}
\value{
A modified Anchovy object, with the recruitment residuals set in R_residuals.
}
\description{
Loads two sets of observed anchovy recruitment residuals into the anchovy
object (for spawning stock biomass below or above 500,000 tons).
}
\keyword{internal}

