% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{make_4panel}
\alias{make_4panel}
\title{Make a 4-panel plot}
\usage{
make_4panel(prey, predator, scenario, nSteps)
}
\arguments{
\item{prey}{A list containing the Prey objects}

\item{predator}{A list containing the Predator objects}
}
\value{
A single 4-panel plot showing the abundance of 4 key species
}
\description{
Make a 4-panel plot that shows the abundance of the key species (spawning
biomass for prey; mature adults for predators) .
}
\keyword{internal}

