% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{makeplot}
\alias{makeplot}
\title{Plot a time series (a utility function)}
\usage{
makeplot(pdata, xname, yname, speciesname, ptitle, y_range, series_type)
}
\arguments{
\item{xname}{A display name for the X-axis}

\item{yname}{A display name for the Y-axis}

\item{speciesname}{A display name for the species}

\item{ptitle}{A partial title}

\item{y_range}{Character; "include_0" to include zero, or "auto"}

\item{data}{A data frame in which the X series to be plotted is in column 1.}
}
\value{
A plot
}
\description{
Plot a time series (a utility function)
}
\seealso{
Other results: \code{\link{calc_predicted}},
  \code{\link{compile_results}},
  \code{\link{get_every_nth}},
  \code{\link{plot_fitted_multipanel}},
  \code{\link{plot_fitted}},
  \code{\link{plot_function_only}},
  \code{\link{plot_summed_by_year}}
}
\keyword{internal}

