% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Other_forage-class}
\alias{Other_forage-class}
\title{An S4 class for Other_forage}
\value{
An object of Other_forage class.
}
\description{
Inherits from Prey class.  By default, this species does not use timestep and
it has a fixed recruitment (R_base).
}
\examples{
other<-new("Other_forage")
#' oth<-new("Other_forage")
slotNames(oth) #see list of slots
oth@Zv[,1:10] #show total mortality by age (Zv) for timesteps 1 to 10
}
\seealso{
parent classes \linkS4class{Prey} and \linkS4class{Species} (top-level).
}

