% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Pelican-class}
\alias{Pelican-class}
\title{An S4 class for Pelicans.}
\value{
An object of the Pelican class.
}
\description{
Inherits from the Predator class.
}
\examples{
pel<-new("Pelican")
slotNames(pel) #see list of slots
pel@Zmat[,1:10] #show total mortality by age (Zmat) for timesteps 1 to 10
}
\seealso{
parent classes \linkS4class{Predator} and \linkS4class{Species} (top-level).
}

