% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{plot_every_nth}
\alias{plot_every_nth}
\title{Plot every nth value.}
\usage{
plot_every_nth(nSteps, pdata, xname, yname, ...)
}
\arguments{
\item{nSteps}{The number of timesteps per year}

\item{xname}{The name of the x series}

\item{data}{A small data frame with the by-timestep series in it}

\item{item}{The name of the time series (matches a colname in data)}

\item{speciesname}{The name of the species}
}
\value{
A plot of every nth value of a vector
}
\description{
A utility function for plotting only the first timestep in each year.
}
\seealso{
Other results: \code{\link{calc_predicted}},
  \code{\link{compile_results}}, \code{\link{makeplot}},
  \code{\link{plot_fitted}},
  \code{\link{plot_function_only}},
  \code{\link{plot_summed_by_year}}
}
\keyword{internal}

