% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot a fitted function (observed vs. expected)}
\usage{
plot_fitted(fitted_data, pdata, xname, yname, speciesname, ptitle, ...)
}
\arguments{
\item{fitted_data}{A dataframe of expected data}

\item{pdata}{A dataframe of observed data}

\item{xname}{X-axis label}

\item{yname}{Y-axis label}

\item{speciesname}{Species name (for title)}

\item{ptitle}{Partial title}
}
\value{
A plot
}
\description{
Plot a fitted function (observed vs. expected)
}
\seealso{
Other results: \code{\link{calc_predicted}},
  \code{\link{compile_results}},
  \code{\link{get_every_nth}}, \code{\link{makeplot}},
  \code{\link{plot_fitted_multipanel}},
  \code{\link{plot_function_only}},
  \code{\link{plot_summed_by_year}}
}
\keyword{internal}

