% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{plot_fitted_multipanel}
\alias{plot_fitted_multipanel}
\title{Plot a multi-panel fitted function (observed vs. expected)}
\usage{
plot_fitted_multipanel(pdata, xname, yname, speciesname, ptitle, ...)
}
\arguments{
\item{xname}{X-axis label}

\item{yname}{Y-axis label}

\item{speciesname}{Species name (for title)}

\item{ptitle}{Partial title}

\item{observed.df}{A dataframe of observed data}

\item{predicted.df}{A dataframe of predicted data}
}
\value{
A plot
}
\description{
Plot a multi-panel fitted function (observed vs. expected)
}
\seealso{
Other results: \code{\link{calc_predicted}},
  \code{\link{compile_results}},
  \code{\link{get_every_nth}}, \code{\link{makeplot}},
  \code{\link{plot_fitted}},
  \code{\link{plot_function_only}},
  \code{\link{plot_summed_by_year}}
}
\keyword{internal}

