% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{plot_function_only}
\alias{plot_function_only}
\title{Plot a function (expected values only)}
\usage{
plot_function_only(fndata, xname, yname, speciesname, ptitle)
}
\arguments{
\item{fndata}{A dataframe of expected data}

\item{xname}{X-axis label}

\item{yname}{Y-axis label}

\item{speciesname}{Species name (for title)}

\item{ptitle}{Partial title}
}
\value{
A plot
}
\description{
Plot a function (expected values only)
}
\seealso{
Other results: \code{\link{calc_predicted}},
  \code{\link{compile_results}}, \code{\link{makeplot}},
  \code{\link{plot_every_nth}}, \code{\link{plot_fitted}},
  \code{\link{plot_summed_by_year}}
}
\keyword{internal}

