% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{plot_summed_by_year}
\alias{plot_summed_by_year}
\title{Sum a by-timestep series by years and plot it.}
\usage{
plot_summed_by_year(nYears, nSteps, pdata, xname, yname, ...)
}
\arguments{
\item{nYears}{The number of years in the time series}

\item{nSteps}{The number of timesteps per year}

\item{xname}{The name of the x series}

\item{data}{A small data frame with the by-timestep series in it}

\item{item}{The name of the time series (matches a colname in data)}

\item{speciesname}{The name of the species}
}
\value{
A plot of values of a vector, summed by year.
}
\description{
A utility function for plotting.
}
\seealso{
Other results: \code{\link{calc_predicted}},
  \code{\link{compile_results}}, \code{\link{makeplot}},
  \code{\link{plot_every_nth}}, \code{\link{plot_fitted}},
  \code{\link{plot_function_only}}
}
\keyword{internal}

