% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Predator-class}
\alias{Predator-class}
\title{An S4 parent class for Predators.}
\value{
Should not be instantiated; rather, instantiate the child classes.
}
\description{
Inherits from Species and has slots for values that are common to predators but not prey.
}
\section{Slots}{

\describe{
\item{\code{sp_prefs}}{list. Prey species preferences (one list element per prey species).}

\item{\code{age_prefs}}{list. A vector of predator preferences for prey age classes; one list element for each prey species.}

\item{\code{omegas}}{list. Parameters that scale available prey biomass per species so they sum (roughly) to 1}

\item{\code{prey_effect}}{list. Parameters for the model that determines how much of an impact prey biomass has on predator recruitment}

\item{\code{Dy}}{numeric. Available, scaled total prey biomass; a vector (time).}

\item{\code{BHp}}{numeric. Beverton-Holt p parameter, which is affected by prey; a vector (time).}

\item{\code{Mo}}{numeric. Maximum number of offspring; a scalar.}

\item{\code{K}}{numeric. Carrying capacity; a scalar.}

\item{\code{Nm}}{numeric. Number of mature animals; a vector.}
}}
\seealso{
parent class \linkS4class{Species}
}

