% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Prey-class}
\alias{Prey-class}
\title{A S4 parent class for Prey.}
\value{
Should not be instantiated; rather, instantiate the child classes.
}
\description{
Inherits from the Species class, and has slots for values that are common to
prey, but not predators.
}
\section{Slots}{

\describe{
\item{\code{B_sp}}{numeric. Spawning biomass by time; a vector.}

\item{\code{B1P}}{numeric. Age 1+ biomass by time; a vector.}

\item{\code{B1PperB0}}{numeric.  Age 1+ biomass relative to unfished age-1+ biomass; a vector (time).}

\item{\code{B0}}{numeric. Average age-1+ biomass in unfished condition.}

\item{\code{G}}{numeric. Environmental driver for reproduction.}
}}
\seealso{
parent class \linkS4class{Species}
}

