% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Sardine-class}
\alias{Sardine-class}
\title{An S4 class for Sardine}
\value{
An object of Sardine class.
}
\description{
Inherits from Prey class.  By default, this species is fished.
}
\examples{
sar<-new("Sardine")
slotNames(sar) #see list of slots
sar@G #Get values of environmental driver, G.
}
\seealso{
parent classes \linkS4class{Prey} and \linkS4class{Species} (top-level).
}

