% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Sealion-class}
\alias{Sealion-class}
\title{An S4 class for Sea lions}
\value{
An object of Sealion class.
}
\description{
Inherits from Predator class.
}
\examples{
lion<-new("Sealion")
slotNames(lion) #see list of slots
lion@Zv[,1:10] #show total mortality by age (Zv) for timesteps 1 to 10
}
\seealso{
parent classes \linkS4class{Predator} and \linkS4class{Species} (top-level).
}

