% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{set_lesliematrix}
\alias{set_lesliematrix}
\title{Put survival rates into a Leslie matrix}
\usage{
set_lesliematrix(P, Zv)
}
\arguments{
\item{P}{A scalar: Maximum age (plus-group)}

\item{Zv}{A vector: Total mortality rate (log scale) at age.}
}
\value{
A leslie matrix
}
\description{
A utility function used in the calculation of numbers-at-age at time t.
Recruits are calculated separately, so fecundities are all set to 0 here.
Survival rates are in off-diagonal rows (survival from age 1 to age 2 is in
leslie[1,2], etc.). Note: age classes are 0:P, but vector and matrix row/col
indexes are 1-based, and go from 1:(P+1=nAges).
}
\keyword{internal}

