% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\name{set_simParam}
\alias{set_simParam}
\title{Set parameter values in a simulationParams object.}
\usage{
set_simParam(params, inputdir, scenario, nYears, nSteps)
}
\arguments{
\item{params}{A simulationParams object.}

\item{scenario}{Character string; a scenario name (must match the name in input_params.csv)}

\item{nYears}{Integer.  Number of years to simulate.}

\item{nSteps}{Integer.  Number of timesteps per year.}

\item{...}{Other arguments.}
}
\value{
A modified simParams object, with values filled from the file.
}
\description{
Reads two files (input_params.csv, which holds most of the input parameters,
and diet_prefs.csv, which holds predator by species dietary preferences) into
two data frames, and loads the data frames into an object of class simulationParams.
}
\seealso{
\linkS4class{simulationParams}

Other initialization: \code{\link{calc_omegas}},
  \code{\link{distribute_byage}},
  \code{\link{distribute_bytimestep}},
  \code{\link{initialize_N}}

Other parameter setting: \code{\link{get_params}},
  \code{\link{get_prefs}}, \code{\link{initialize_N}},
  \code{\link{set_species}}
}
\keyword{internal.}

