% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{methods}
\name{set_species}
\alias{set_species}
\alias{set_species,Predator,simulationParams-method}
\alias{set_species,Species,simulationParams-method}
\title{Set initial parameter values in Species objects.}
\usage{
set_species(species, params, ...)

\S4method{set_species}{Species,simulationParams}(species, params, ...)

\S4method{set_species}{Predator,simulationParams}(species, params, ...)
}
\arguments{
\item{...}{Other arguments (from child methods).}

\item{species.}{A Species object.}

\item{params.}{A simParams object.}
}
\value{
A child method, depending on signature.
}
\description{
Set initial parameter values in Species objects.

The default method sets parameter values in a Species object.  It calls
functions to read in parameters from input files, and to initialize arrays.

The method for predators also reads in a predator's diet preferences (for
prey species and the age-classes of those prey).
}
\section{Methods (by class)}{
\itemize{
\item \code{species = Species,params = simulationParams}: Default method.

\item \code{species = Predator,params = simulationParams}: Also reads in diet preferences.
}}
\examples{
speciesA<-set_species(speciesA)
}
\seealso{
Other parameter setting: \code{\link{get_params}},
  \code{\link{get_prefs}}, \code{\link{initialize_N}},
  \code{\link{set_simParam}}
}
\keyword{internal}

