% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{simulationParams-class}
\alias{simulationParams-class}
\title{An S4 class to hold global parameters and input data for simulations.}
\value{
A simulationParams object.
}
\description{
Species-specific parameter values are held in two dataframes.  Global values
have their own slots.  This class is used to read input data from two external
files, and is later used to initialize Species objects with species-specific
parameter values.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. The object name (should match the name of the class).}

\item{\code{nYears}}{numeric. Number of years for the simulation.}

\item{\code{nSteps}}{numeric. Number of timesteps per year (may not be used by all species)}

\item{\code{scenario}}{character. A name for the scenario, which must match a column name in input_params.csv.}

\item{\code{params}}{A dataframe that holds parameter values for everything except diet preferences.}

\item{\code{diet_prefs}}{A dataframe that holds diet preferences for predators.}
}}

