% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_initialization.R
\name{solve_age0_survival}
\alias{solve_age0_survival}
\title{Solve for age-0 survival}
\usage{
solve_age0_survival(predator)
}
\arguments{
\item{predator}{A Predator object.}
}
\value{
A predator with modified Mv slot (set age-0 survival, Mv[1]).
}
\description{
Solves for age-0 survival of predators iteratively, by balancing reproduction
and mortality (ignores prey effect, so this is unfished prey condition).
}
\seealso{
Other initialization: \code{\link{calibrate}},
  \code{\link{distribute_byage}},
  \code{\link{distribute_bytimestep}},
  \code{\link{initialize_N}}, \code{\link{set_simParam}}
}
\keyword{internal}

