% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_initialization.R
\name{solve_fecundity}
\alias{solve_fecundity}
\title{Solve for the scaling parameter for fecundity.}
\usage{
solve_fecundity(predator)
}
\arguments{
\item{predator}{A Predator object.}
}
\value{
A modified Predator object with bigPhi set in the prey_effect slot.
}
\description{
Finds the value of a scaling parameter for predator reproduction (upper-case
Phi in Punt et al. 2016, equation 9 line 1).  Note: you must solve for age-0
survival before calling this function.
}
\keyword{internal}

