% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{solve_for_Ft}
\alias{solve_for_Ft}
\title{Solve for F}
\usage{
solve_for_Ft(C, Nv, wv, Mv, Sv, Fstart)
}
\arguments{
\item{C}{Target catch.}

\item{Nv}{Number at age vector.}

\item{wv}{Weight at age vector.}

\item{Mv}{Mortality at age vector.}

\item{Sv}{Selectivity at age vector.}

\item{Fstart}{Starting guess at F.}
}
\value{
Instantaneous fishing mortality rate, Ft (a scalar).
}
\description{
Iteratively solve for instantaneous fishing mortality, F, given some target
catch.  Assumes the target is between 0 and 2x the initial guess.
}
\keyword{internal}

