% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Species-class}
\alias{Species-class}
\title{An S4 class to represent a generic species.}
\value{
Should not be instantiated; rather, instantiate the child classes.
}
\description{
It holds values that are used to track the population growth (including
reproduction, survival, and exploitation) of every species, and all other
species objects inherit from this class. The variable names that end in 'v'
are vectors; the names that end in 'mat' are matrices.  The rest are scalars,
except for a few that (as shown) are lists.  Lists are used to hold
parameters for a model.  Depending on the model, those parameters may be
scalars, vectors, or a combination.  Holding model parameters in lists makes
it easier to swap models later.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Species name.  Should match the class name.}

\item{\code{uses_timestep}}{logical. Is the species tracked by timestep within a year (or just by year)?}

\item{\code{P}}{numeric. Age plus-group; a scalar.}

\item{\code{am}}{numeric. Age at maturity; a vector.}

\item{\code{mv}}{numeric. Proportion mature at age; a vector.}

\item{\code{Ninit}}{numeric. Numbers at age in year 1 (or timestep 1 of year 1); a vector.}

\item{\code{Nmat}}{matrix.  Population matrix, age x time (years, or timesteps x years).}

\item{\code{wv}}{numeric. Weight at age; a vector.}

\item{\code{M}}{numeric. Natural mortality rate (log scale); a scalar.}

\item{\code{fished}}{logical.  Is the species fished?}

\item{\code{fishing_rule}}{list. Parameters for the fishing rule model; a list.}

\item{\code{Sv}}{numeric. Fishing selectivity at age; a vector.}

\item{\code{Fv}}{numeric. Fishing mortality rate (log scale) by time; a vector.}

\item{\code{F_baseline}}{numeric. Baseline fishing mortality rate; a scalar.}

\item{\code{gv}}{numeric. Proportion of F that occurs per timestep; a vector.}

\item{\code{Cv}}{numeric. Catch by time; a vector.}

\item{\code{Zv}}{matrix. Total mortality rate (log scale) by time: a vector.}

\item{\code{R_baseyear}}{numeric. Recruitment in a year 1, if supplied; a scalar.}

\item{\code{R}}{numeric. Recruitment by year; a vector.}

\item{\code{lnSR_err}}{numeric. Stock-recruitment errors (ln scale) by time; a vector.}

\item{\code{recruitment}}{list. Parameters used in recruitment model (may vary by species)}

\item{\code{B1P}}{numeric. Age 1+ biomass by time; a vector.}
}}

