% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_functions.R
\name{calc_catch}
\alias{calc_catch}
\title{Calculate catch (tons) for a timestep or year}
\usage{
calc_catch(Species, t)
}
\arguments{
\item{Species}{A Species object}

\item{t}{time period (may be year or global timestep)}
}
\value{
A Species object with catch set if fished=TRUE, or unset if fished=FALSE
}
\description{
Calculate catch (tons) for a timestep or year
}
\seealso{
Other fishing: \code{\link{calc_F_t}},
  \code{\link{calc_F_y}}
}
\keyword{internal}

