% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_results.R
\name{compile_results}
\alias{compile_results}
\title{Compile results and generate graphs.}
\usage{
compile_results(species, plot_master, nYears, nSteps)
}
\arguments{
\item{species}{A Species object}

\item{nYears}{Number of years in the simulation.}

\item{nSteps}{Number of timesteps per year (ignored for some species)}

\item{slotlist}{A list of slots, from which graphs are to be made.}
}
\value{
A list with the values of certain vectors assembled into a data frame, and a list of plots.
}
\description{
Creates graphs of objects in Species slots, taking instructions from the
plotlist.csv file.  Note that since a dataframe cannot be a ragged array, the
number of rows in the data frame returned by the function is either nYears or
tTimes, but certain items (catch, recruits) are plotted by year, even for
species that are tracked by timestep.
}
\seealso{
Other results: \code{\link{calc_predicted}},
  \code{\link{makeplot}}, \code{\link{plot_every_nth}},
  \code{\link{plot_fitted}},
  \code{\link{plot_function_only}},
  \code{\link{plot_summed_by_year}}
}
\keyword{internal}

