% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_initialization.R
\name{distribute_bytimestep}
\alias{distribute_bytimestep}
\title{Distribute numbers-at-age by timestep}
\usage{
distribute_bytimestep(Nv, Mv, L)
}
\arguments{
\item{Nv}{Population size by age in timestep 1; a vector.}

\item{Mv}{Natural mortality rate (log scale) by age; a vector.}

\item{L}{Number of timesteps per year; scalar.}
}
\value{
A small matrix of population size in year 1; dim=(nAges x L).
}
\description{
#Take an initial vector of population by age for a species and distribute it
across all timesteps within one year.  Apply cumulative natural mortality to each
step.
}
\seealso{
Other initialization: \code{\link{calibrate}},
  \code{\link{distribute_byage}},
  \code{\link{initialize_N}}, \code{\link{set_simParam}},
  \code{\link{solve_age0_survival}}
}
\keyword{internal}

