% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE_object_model.R
\docType{class}
\name{Prey-class}
\alias{Prey-class}
\title{A S4 parent class for Prey.}
\value{
Should not be instantiated; rather, instantiate the child classes.
}
\description{
Inherits from the Species class, and has slots for values that are common to
prey, but not predators.
}
\section{Slots}{

\describe{
\item{\code{availability}}{list. Parameters of the spatial availability function.}

\item{\code{observed_SR_err}}{numeric. Observed stock-recruit residuals (ln scale), to resample from.}
}}
\seealso{
parent class \linkS4class{Species}
}

